+++ contracts/oracles/IUniV2PriceOracle.sol
@@ -5,10 +5,19 @@ pragma solidity ^0.8.0;
 // @title IUniV2PriceOracle - Price oracle
 interface IUniV2PriceOracle {
     function update() external;
+
     function PERIOD() external view returns (uint256);
+
     function blockTimestampLast() external view returns (uint32);
+
     function pair() external view returns (address);
+
     function token0() external view returns (address);
+
     function token1() external view returns (address);
-    function consult(address token, uint amountIn) external view returns (uint amountOut);
+
+    function consult(address token, uint256 amountIn)
+        external
+        view
+        returns (uint256 amountOut);
 }
